+++ src/app/_components/HeaderTabs.tsx
@@ -1,7 +1,7 @@
 'use client';
 import Link from 'next/link';
 import React from 'react';
-import sty from './css/header.module.css';
+import sty from './header.module.css';
 
 
 export default function HeaderTabs() {
