+++ src/app/page.tsx
@@ -1,5 +1,5 @@
 import Header from './_components/Header';
-import sty from './_components/css/main.module.css'
+import sty from './_components/main.module.css'
 
 export default function Home() {
   return <div className={sty.main_wrapper}>
